Nerf Chem warrior vs. Wood, 50% -> 35%
Since Chem warriors are awesome vs. everything, I think it would make sense to make them not quite as awesome vs. buildings. There is a bit of a lack of dinstinction between chem and flamers, and right now it doesn't appear to players like it makes much sense to build flamers when you can build chems instead. This will keep "burning buildings down" as the forte of the flame throwers. 
The chems will still be decent vs. wood, but not as crazy as they have been.