Reduce grenadier vs. wood 75% -> 50%
Grenadiers actually do more damage than flamethrowers vs. buildings, which is counter-intuitive and not in keeping with flavour and design intent. They should be good vs. buildings, but not quite this good, especially given that they are cheaper than flamethrowers. This nerf still allows them to be effective at taking out buildings, but not quite as blindingly fast at destroying them.