Fixes #5654 by removing false damage states which caused invisible walls
Removes final set of 16 sequences for all walls as it appears that these
are not actually supposed to be used as damage states for the actual
wall, but as end pieces for each wall.  One of the pieces in particular
caused a single wall to be seen as invisible.  Fixes #5654